Added champion.gg API and !counter; removed RiotAPI.py
This commit adds champggapi.py, which is an API for champion.gg (duhh).
It also adds !counter, which is in league.py and scrapes champion.gg to
find counters for a champion of choice. Syntax is: !counter <champion>
<lane> <number of counters to show>

Also, I removed RiotAPI.py because it isn't being used any more.