And thus begins our use of champion mastery
Added get_top_champ_mastery() to rawpi.py. It does... exactly what you
think it does.

Added an early version of !mastery to league.py. It gets the top 3
champs a summoner has mastery on and spits it out in chat in dictionary
form. 3 is a magic number but I'll change it into an argument in the
next version.

The code is kind of a mess though. Also, !mastery can handle one API
exception but not two for some reason. aafahfeijf I'm too sleepy to keep
going but it works and i'm excited